Converted metrica_revisions to .tex
There were some issues with knitr and synctex so I deleted the various
files related to and including metrica_revisions.Rnw and created a .tex
version of this file. The R code for the figure is commented out inside
the .tex file. Eventually I'll transfer this to its own .R file and
read in the output of TikzDevice.