Completed AMSE-Optimal Weighting Material
Finished the derivations of the (infeasible) AMSE-optimal weights for
OLS and 2SLS. Corrected a sign error and checked that these results do
in fact agree with my earlier derivations for the special case in which
there is only one instrument. Added a subsection in which I will work
out the details for implementing a feasible version of the AMSE-optimal
weights.