Implementing FMSC via Hausman Test
Filled in some details and simplified this material. Explains how we
can actually implement the FMSC for the OLS vs 2SLS example via a
Hausman test, which quantities we need to estimate, etc. Also added
material discussing the potential problem of the difference of
asymptotic variance estimators not being positive. This can be avoided
by using the same estimator of the error variance for each of the two
pieces.