Renamed pmse.compare, deleted non-parallel version
I eliminated the non-parallel function mse.compare and renamed
pmse.compare. The latter is much faster when the user has multiple
cores but it will still run on machines with a single core. The package
multicore is only available for Linux/Mac, however, which I should
mention in the README.