Test mse.compare against mse.compare.cpp
I wrote a few lines of code to test the two versions of mse.compare
against each other: one using C++ code for the dgp and one using R
code. The substantive results were very similar, but the C++ code
produced much "smoother" graphs. Is this because it is using the same
random seed for each of the different parameter values? I need to
figure this out...