Created and tested dgp.alt
This is an alternative version of my dgp function that doesn't use the
library MASS. Instead it creates multivariate normals directly from
rnorm using the Cholesky Decomposition. I created this function for
eventual comparison to a C++ version that will use the Rcpp version of
rnorm. I want to be able to make sure I'm getting exactly the same
random numbers as I would get in R, set the seed, etc.