Tests of dgp_cpp_alt against dgp
The two yield identical results when I set the same seed, showing that
I am indeed generating the same random draws. It took me a while to see
this because mclapply was messing things up in the way it set the seed
for different processes. I temporarily switched back to plain old
lapply and this resolved the problem.