More tests, deploy mclapply in the final step
Ran some more tests to make sure that I'm getting the same random
numbers from Rcpp when I set the same seed and that these agree with
the R random numbers. Added mclapply in the final step where parameter
values are varied. There's an option that tells it to use the same seed
for each process. This works as expected.