Constrained weight to lie in (0,1)
I thought this was necessary since I was getting the strange result
that the averaging estimator had a lower RMSE than either the OLS or
2SLS estimators. At first I thought this must be a coding error, but I
went through and extracted the weights and reconstructed the weighting
estimator and its RMSE "by hand" evaluated at the offending parameter
values and everything checked out. I realize now that this is not a
contradiction. First of all, the weight is random here, not fixed. Even
if the weight were a constant, the variance of a convex combination of
RVs can be lower than the convex combination of the variances! This
should have been obvious but somehow I missed it...