First working version of fmsc_ols_iv_cpp
Wrote and tested simplest C++ version of the fmsc function for the OLS
versus IV example. Works as expected and gives the same numeric result
at the R version. As of now, the C++ version doesn't calculate the
averging estimator or the DHW pre-test estimators. Also it returns a
list rather than a vector. The next step is to address these
differences.