slight change in code for calculating AMSE optimal weight
Changed the name of a variable to sq_bias_est and removed the
unncessary code that constrains the weights to lie in [0,1]. They are
guaranteed to be interior as long as the squared bias estimate can't be
negative, which we impose by taking the max of our estimate and zero.
Also corrected a typo inside an if statement of the R version that
never caused any trouble before since the condition was never
satisfied! I tested the changes by first making them to the C++ code
and comparing the results to the unchanged R code and finally changing
the R code and comparing to the changed C++ code. Everything works as
expected.