renamed s_e_sq
Since I’m going to have both an estimator of the variance of epsilon
based on the tsls and old residuals in my code, I need names to match.
Up until this point I’ve only ever needed the version based on the tsls
residuals so I didn’t make this clear in the name. Now I’ve explicitly
added the “_tsls” suffix. I tested everything and it works as expected.