Additions to fmsc_OLS_IV class
Created two new private data members, ols_resid and s_e_sq_ols, and
added code to the constructor to calculate them. These are the
residuals based on the old fit and the corresponding estimate of the
variance of epsilon, respectively. I’ll need these for the naive and
old-based confidence intervals that I’ll implement next.