changed constructor to pass by constant reference
There’s no reason to make copies of the vectors and matrices and
they’re not going to be modified. I thought this would speed things up,
but it doesn’t seem to have made much of a difference. The results are
still the same though and I’ve gotten some practice with a new C++
concept!