Textbook SE now works for C++
I made a bunch of member functions for tsls_fit public, simplified the
code to calculate Rtilde_inv using the fact that the inv function in
arms understands trimatu, and corrected an error in the calculation of
s_e: I was trying to calculate n from the length of the residuals
before I had actually calculated them! Now it works and gives the same
result as tsls in the sam library. I had to divide by n - k rather than
n in the calculation of s_sq to get it exact since this is what tsls
does.