additions to constructor
I'll need the constructor to calculate the matrices K and Omega for
each of the candidates. Started working on code that will accomplish
this. The matrix candidates, if specified, is an indicator matrix
corresponding to the columns of the z2 that specifies additional
estimators besides the full and valid estimators. If unspecified, we'll
just do the calculations for the full and valid. Otherwise we'll loop
over candidates to construct the K and Omega matrices.