created a simple example of function pointers
The function ResultPlusOne doesn't need to "know" exactly how a and b
are turned into a result. It only knows what type everything is and
that it should add one to whatever result it gets by calling the
function that is passed to it. This is a prototype of how I will
implement the focus parameter in my FMSC code.