substantially simplified the constructor
There was a lot of repetition in the if else branch originally which
I've now removed. It compiles but I still need to test it. Next step is
to move the calculation of the inner bias matrix and variance matrix
for each candidate into the constructor and out of the other member
functions. It should go into the for loop.