Simple test example for subletting matrices in armadillo
I didn't initially understand how matrix subletting worked in
armadillo. I thought you could pass a uvec of ones and zeros of the
same length as the number of columns and use this to get the desired
subset of columns but this is not correct. When we use .cols(uvec
something) the uvec must contain the column indices. The solution is to
first use find to convert a vector of ones and zeros into a vector
containing the indices of the non-zero elements.