correct matrix sub-setting syntax error
This commit corrects the fmsc_chooseIV class in the same way that
commit 95b148e corrected the linearGMM_select class. Unlike R,
Armadillo does not us to use a vector of zeros and ones to subset
matrices. By using the function find() we can convert such a vector of
zeros and ones into a vector of the *positions* of nonzero elements.
The result can then be used to extract submatrices, etc.

I recompiled the code at this point and there were no syntax errors. I
still need to test it though.