correct matrix sub-setting error
This is a follow-on to the correction from commit 1feb593. When I
initially wrote this code I didn't completely understand Armadillo's
syntax for matrix sub-setting and erroneously tried to use vectors of
zeros and ones to select columns by analogy with R. This is not in fact
valid Armadillo syntax. The solution is to use find() to convert a
vector of ones and zeros to a vector of the positions of the ones.