fix DGP class and references to it
There was a *major* error in my DGP class: the transformation from
independent standard normals to correlated standard normals was
*incorrect* so I wasn't getting the draws I thought I was. To prevent
errors of this kind in the future, I wrote the C++ function mvrnorm
mimicking the functionality of the function of the same name from the
MASS function in R. Now the DGP class calls this function instead of
doing the operations in place.

I also renamed the variables to match the paper: u becomes epsilon,
epsilon becomes v, and z2 becomes w. Since there's no longer a z2
floating around, z1 becomes z.

Since the variables in the DGP class changed, I also updated the
references to it in mse_compare_cpp.