update RMSE plotting code
Because the output of the script that calculates MSE for the OLS vs IV
simulation was changed in the preceding commit to be a data frame
rather than a matrix, I had to update this script as well.

The basic structure is the same. Essentially, I "repackage" the results
of the MSE calculations for compatibility with the existing plotting
code.

There also a few minor tweaks here and there but the only important one
is that the user must now specify the layout of the panels in the plot
by providing the number of rows and columns. I've added this as a
"macro" at the top of the file to make it easy to change.