change "pi" to "gamma"
In the OLS versus IV simulation, we varied Pi, the strength of the
first stage. In this simulation, two parameters govern the strength of
the first stage: Pi and Gamma. We hold gamma fixed in the simulation
and vary Gamma. Hence, Gamma here plays the role of Pi in the OLS vs IV
simulation.