redefine K to match notation in paper
In the present version of the paper, I am completely consistent in my
definition of K for both the general case and the examples. In my
earlier notes, however, I used a different convention for the sign of K
in the choosing IVs example.

I've now changed this. I've also added some clarifying remarks about
the matrix Psi and how it relates to the chosen sign convention for K.