change sign convention for K to match paper
This commit changes my C++ code in the same way that the preceding
commit changed my notes.

I formerly used a different sign convention for K in the choosing IVs
example than I used in the general GMM case. This frequently confused
me, particularly as it related to the matrix Psi.

I've now updated everything so the same sign convention for K is used
throughout.