correct error in definition of Psi for IV example
This is a follow-on to the previous two commits. When I changed the
sign convention for K in the choosing IVs example from the paper, I
forgot that I had been using a different convention in my notes. This
created an inconsistency in my definition of Psi. This commit corrects
the mistake.