correct CRUCIAL error in J-test p-value
A typo messed up my p-value calculations for the J-test! We need the
*upper tail* of the chi-squared distribution but I had the arguments
backwards: 1 means lower.tail = TRUE and 0 means lower.tail = FALSE.
This explains why both J90 and J95 looked the same and were identical
to the full model! I'll need to re-run the simulation to implement this
change.